# [SPARK-16186][SQL] Support partition batch pruning with `IN` predicate in InMemoryTableScanExec