@dongjoon-hyun 2016-06-24
Hi, @cloud-fan .
I updated the PR. IMO,
- InSet is used for large size of `IN` .
- This PR is used for small size of `IN`.