@dongjoon-hyun 2016-06-24
Although I decided to make this PR after observing TPC-DS queries, I will definitely update this PR if there are another useful scenarios.