@cloud-fan 2016-06-24
Seems like a reasonable optimization to me, cc @liancheng