@davies 2016-06-24
@dongjoon-hyun Thanks for the patch, this optimization sounds reasonable.

I'm thinking of is it possible to make the optimization for IN/INSET more general. We could have a optimizer to insert a GreaterThanOrEqual and LessThanOrEqual for IN/INSET (checking the data type to make sure they are orderable), it will be cheaper to evaluate. For IN with single expression, we could rewrite as EqualTo. When doing this, we should be careful for null, to respect the sematics of null in IN/INSET. By doing this, all the data sources could benefit from it.