@davies 2016-06-24
BTW, we could use constraints to implement this.