@davies 2016-06-24
For any IN that have more than one expression, we could add another GreaterThanOrEqual/LessThanOrEqual (not replace the IN).

For 2, it's not that obvious yet, we can do that later.