@dongjoon-hyun 2016-06-24
Oh, you meant adding additional constraints by using **min** and **max**. I see.

By the way, I have one question. If there are many predicates, does Spark use the predicate in a sorted order?

I'm not sure the newly inserted `GreaterThanOrEqual/LessThanOrEqual` can be used first.