@dongjoon-hyun 2016-06-24
Okay. Let's summarize before updating PR.
1. In general, `a IN (expression)` will be `a = expression`. `OptimizerIn` optimizer will take care of this.
2. In general, `a IN (2001, 2002, ..., 2009)` will be `2001 <= a AND a <= 2003 AND a IN (2001, 2002, 2009)`. `OptimizerIn` optimizer will take care of this.
Am I correct?