@davies 2016-06-24
@dongjoon-hyun Yes, 2) should check the constraints to make it idempotent