@dongjoon-hyun 2016-06-24
Hmm. The general idea is good. But, I still think this PR and the idea seem to be complementary to each other.

Sorry, but, if possible, can we proceed that general idea in another PR? 

It's a little bit beyond the title of this PR and the touched files will be different.