@dongjoon-hyun 2016-06-24
The some of frequent TPC-DS usages were STATE, ZIP, Color strings. The min/max of these values doesn't have much meaning.