@davies 2016-06-24
@dongjoon-hyun That's a good point, the current patch is better for performance actually