@davies 2016-06-24
Let's go with current patch, I will review it now. Those things could be considered later.