@dongjoon-hyun 2016-06-25
Hi, @davies .
I removes the option-related stuff from the code/PR description/JIRA description according to your advice.
Thank you for review!