@davies 2016-06-25
LGTM, merging this into master.