@davies 2016-06-25
Reverted, will merge this again once it passed jenkins.