@dongjoon-hyun 2016-06-25
Maybe, you are confused with https://github.com/apache/spark/pull/13900 .
It passed.