@dongjoon-hyun 2016-06-25
Hi, @davies .
Now, it passed.
If there is anything for me to do, please let me know.