@davies 2016-06-25
Merged into master, thanks